+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -10,13 +10,13 @@ import javax.validation.constraints.NotBlank;
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class UserDto {
-    Integer id;
+    Long id;
     String name;
     @Email(message = "Невалидная почта")
     @NotBlank(message = "Почта не может быть пустой")
     String email;
 
-    public UserDto(Integer id, String name, String email) {
+    public UserDto(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
