+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -10,7 +10,7 @@ import javax.validation.constraints.NotBlank;
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class User {
-    Integer id;
+    Long id;
     String name;
     @Email(message = "Невалидная почта")
     @NotBlank(message = "Почта не может быть пустой")
@@ -21,7 +21,7 @@ public class User {
         this.email = email;
     }
 
-    public User(Integer id, String name, String email) {
+    public User(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
